requirements.txt
Django
djangorestframework
twine
wheel
drf-simple-invite
factory_boy